chore: fixes
Simplify code in a manner after God's own heart.
Or maybe just mine.
This fixes a lot of type holes.
We now use template literal types, custom type literal functions, and don't use casts.
In addition, a lot more things are `readonly` or `const` now. Down with mutability!

UI-wise, I removed the link to the homepage in the header, because you can just click the "logo."
This simplifies quite a bit of code.