chore(deps): bump
Bump preact.
Downgrade gfm to a csr compatible version.

HeadlessUI v2.0 is broken, and HeadlessUI v1.18 added code from v2.0.
It's mostly turned off, but it's still in the bundle.
It relies on process.env.JEST_WORKER_ID to disable some features.
We use combobox, the affected component, so we can't upgrade.