fix: header animations
6c9ba3f3520c6146782a3a449a6a854aaafaba46 broke them.
It was a step in the right direction.
However, it actually made it worse in the short run.
You see, `animation` has to be applied before `animation-delay`.
By fixing the selector, it fixed the specificity, which revealed this.
This moves the animation delays to the end of the file.
That makes sure that they're applied after the animations.

Hopefully, this is the last time I have to hunting this stuff down for a while.
Flutter for the win!